[#174624167] Add timeout to fetch + refactor
